+++ backend/src/main/java/site/coduo/member/controller/dto/oauth/State.java
@@ -5,7 +5,7 @@ import site.coduo.member.exception.AuthenticationException;
 public record State(String value) {
 
     public void validate(State other) {
-        if (!value.equals(other.value)) {
+        if (!(value.equals(other.value))) {
             throw new AuthenticationException("state값 불일치!");
         }
     }
